+++ environments/common_env.rb
@@ -1,2 +1,4 @@
 require_relative '../lib/cuke_modeler'
 require_relative '../testing/file_helper'
+
+MOST_CURRENT_GHERKIN_VERSION = ENV['MOST_CURRENT_GHERKIN_VERSION'].to_i