+++ lib/cuke_modeler/parsing.rb
@@ -7,7 +7,7 @@ require 'gherkin'
 # an 'adapter' appropriate to the version of the *cucumber-gherkin* gem that has been activated.
 gherkin_version = Gem.loaded_specs['cucumber-gherkin'].version.version
 gherkin_major_version = gherkin_version.match(/^(\d+)\./)[1].to_i
-supported_gherkin_versions = (9..19)
+supported_gherkin_versions = (9..20)
 
 raise("Unknown Gherkin version: '#{gherkin_version}'") unless supported_gherkin_versions.include?(gherkin_major_version)
 
@@ -61,6 +61,23 @@ module CukeModeler
       # inside of it, so I'm leaving this here in case it changes again
       # rubocop:disable Lint/DuplicateMethods
       case gherkin_major_version
+        when 20
+          # TODO: make these methods private?
+          # NOT A PART OF THE PUBLIC API
+          # The method to use for parsing Gherkin text
+          def parsing_method(source_text, filename)
+            messages = Gherkin.from_source(filename,
+                                           source_text,
+                                           { include_gherkin_document: true })
+                              .to_a
+
+            error_message = messages.find { |message| message.parse_error }
+            gherkin_ast_message = messages.find { |message| message.gherkin_document }
+
+            raise error_message.parse_error.message if error_message
+
+            gherkin_ast_message.gherkin_document
+          end
         when 19
           # TODO: make these methods private?
           # NOT A PART OF THE PUBLIC API