+++ rakefiles/ci_tasks.rb
@@ -1,6 +1,6 @@
 namespace 'cuke_modeler' do
 
-  desc 'The task that CI will run. Do not run locally.'
+  desc 'The task that CI will run.'
   task :ci_build => ['cuke_modeler:full_check'] # rubocop:disable Style/HashSyntax
 
 end